Вынес ApiFactory в dagger модуль и пометил компонент ApplicationScope
