+++ app/src/main/java/com/example/pizza_shift_2025/pizza_screen/di/PizzaModule.kt
@@ -1,7 +1,9 @@
 package com.example.pizza_shift_2025.pizza_screen.di
 
 import com.example.pizza_shift_2025.common.data.remote.ApiFactory
+import com.example.pizza_shift_2025.common.data.remote.ApiFactoryImpl
 import com.example.pizza_shift_2025.pizza_screen.data.PizzasRepositoryImpl
+import com.example.pizza_shift_2025.pizza_screen.data.remote.PizzaService
 import com.example.pizza_shift_2025.pizza_screen.domain.PizzasRepository
 import dagger.Binds
 import dagger.Module
@@ -15,6 +17,8 @@ interface PizzaModule {
 
     companion object {
         @Provides
-        fun providePizzaService() = ApiFactory.pizzaService
+        fun providePizzaService(apiFactory: ApiFactory): PizzaService {
+            return apiFactory.createApi(PizzaService::class.java)
+        }
     }
 }
